Move dynamic test to separate file
